feat: 상품 집계 도메인 서비스 구현

+++ apps/commerce-api/src/test/kotlin/com/loopers/support/fixture/ProductSummaryFixture.kt
@@ -4,7 +4,7 @@ import com.loopers.domain.product.ProductSummary
 
 fun createProductSummary(
     productId: Long,
-    likeCount: Int = 100,
+    likeCount: Long = 100,
 ): ProductSummary = ProductSummary(
     productId = productId,
     likeCount = likeCount,
